+++ src/main/java/pl/edu/wszib/book/store/dao/IUserDAO.java
@@ -6,10 +6,10 @@ import java.util.List;
 import java.util.Optional;
 
 public interface IUserDAO {
-    Optional<User> getById(int id);
+    Optional<User> getById(Long id);
     Optional<User> getByLogin(String login);
     List<User> getAll();
     void save(User user);
-    void remove(int id);
+    void remove(Long id);
     void update(User user);
 }
