+++ src/main/java/pl/edu/wszib/book/store/services/IBookService.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 
 public interface IBookService {
     void save(Book book);
-    Optional<Book> getById(int id);
-    void update(Book book, int id);
+    Optional<Book> getById(Long id);
+    void update(Book book, Long id);
     List<Book> getAll();
     List<Book> getByPattern(String pattern);
 }
