+++ src/main/java/pl/edu/wszib/book/store/services/ICartService.java
@@ -2,5 +2,5 @@ package pl.edu.wszib.book.store.services;
 
 public interface ICartService {
     double calculateCartSum();
-    void addBookToCart(int id);
+    void addBookToCart(Long id);
 }
